+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/integration/SentryPluginAutoInstallTest.kt
@@ -347,9 +347,60 @@ class SentryPluginAutoInstallTest :
         assertFalse { "io.sentry:sentry-android:6.32.0 -> 7.0.0" in result.output }
     }
 
-    private fun runListDependenciesTask() = runner
+    @Test
+    fun `considers sentry-opentelemetry-agentless as a core version`() {
+        appBuildFile.writeText(
+            // language=Groovy
+            """
+            plugins {
+                id "java"
+                id "io.sentry.jvm.gradle"
+            }
+
+            dependencies {
+              implementation 'io.sentry:sentry-opentelemetry-agentless:8.0.0'
+            }
+
+            sentry.autoInstallation.enabled = true
+            sentry.autoInstallation.sentryVersion = "8.1.0"
+            """.trimIndent()
+        )
+
+        val result = runListDependenciesTask(
+            extraArguments = listOf("--configuration", "runtimeClasspath")
+        )
+        assertTrue { "io.sentry:sentry:8.0.0" in result.output }
+    }
+
+    @Test
+    fun `considers sentry-opentelemetry-agentless-spring as a core version`() {
+        appBuildFile.writeText(
+            // language=Groovy
+            """
+            plugins {
+                id "java"
+                id "io.sentry.jvm.gradle"
+            }
+
+            dependencies {
+              implementation 'io.sentry:sentry-opentelemetry-agentless-spring:8.0.0'
+            }
+
+            sentry.autoInstallation.enabled = true
+            sentry.autoInstallation.sentryVersion = "8.1.0"
+            """.trimIndent()
+        )
+
+        val result = runListDependenciesTask(
+            extraArguments = listOf("--configuration", "runtimeClasspath")
+        )
+        assertTrue { "io.sentry:sentry:8.0.0" in result.output }
+    }
+
+    private fun runListDependenciesTask(
+        extraArguments: List<String> = listOf("--configuration", "debugRuntimeClasspath")
+    ) = runner
         .appendArguments("app:dependencies")
-        .appendArguments("--configuration")
-        .appendArguments("debugRuntimeClasspath")
+        .appendArguments(*extraArguments.toTypedArray())
         .build()
 }