+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## Unreleased
 
+### Features
+
+- Allow `sentry-opentelemetry-agentless` and `sentry-opentelemetry-agentless-spring` to determine the SDK version to install ([#841](https://github.com/getsentry/sentry-android-gradle-plugin/pull/841))
+
 ### Fixes
 
+- Ensure `InjectSentryDebugMetaPropertiesTask` deletes the output folder before writing into it ([#838](https://github.com/getsentry/sentry-android-gradle-plugin/pull/838))
 - Avoid excessive sentry-cli search logs at `info` level ([#842](https://github.com/getsentry/sentry-android-gradle-plugin/pull/842))
 
 ### Dependencies